Implemented ugly (but effective) hack to make custom locations work better.
When the custom location is created, the actor can't find their PackageLocation if that cell wasn't loaded, which it usually won't be.
So instead we get the Cell from the JDB, then use GetNthRef to find an object in it that IS loaded.
Nearly all Cells have at least one object that's a Property or is filling an alias, so this gives us a base to send our character to when they're summoned.
We still need a fix for if this fails, and a generally better solution would be preferred, but it works surprisingly well so far.