MCM can now be used to re-sort the Shrine alcoves.
Dumbed down the Alcove script, too, which actually made the UpdateShrineNames function simpler. Two passes are made. The first one empties any shrines that are going to be switched. The second then repopulates them. Now any series of changes to the Alcove characters can be applied with a single function.