Created detailed requirements list in character files.
This ticks serialization version to 3. Not really a compatibility-breaking change, but future versions will rely on the new data being there to notify players of missing mods. Added code to upgrade the files if they are behind based on the current setup. This should be generally safe because not many people are trading files yet.