Alcoves now check for stuck validation and time out after 2 minutes.
For now the timeout just notifies the player of the problem and tells them to go to MCM. Perhaps they should auto-reset, but on slow systems with a full Shrine that might cause characters to be removed while still loading. Closes #72, though: it should now be nearly impossible to get the Shrine stuck in an invalid state unless script latency is intolerably high.