Re-add Perk support. Minor bugfixes.
Doppelgangers like to strip any time they switch weapons, not sure why. For now armor is equipped with the DisallowRemoval flag on in vMYC_Doppelganger.psc.

Thing with the formlists is cumbersome, but I think it's the only way to deal with threading without making a separate formlist for every actor, which would suck.